makesig: Fail if operating on non-contiguous section
The current assumption is that the next instruction is immediately after the previous one.
That doesn't apply in certain situations, such as when bytes are padded / undefined.

The script previously didn't account for this, so the signature would not correspond to
the actual function, if it even did find one.

This closes #1.